Refactor: Separate rendering logic from CurriculumManager
- Extracts all DOM rendering logic into a new `CurriculumRenderer` class.
- Extracts all canvas payload generation logic into a new `CanvasBuilder` class.
- Refactors `CurriculumManager` to delegate rendering and canvas building tasks to the new classes.
- This change improves separation of concerns and makes the `CurriculumManager` class more focused on data management.
